Running with NO_MILLIS caused the timers not to be setup for analogWrite().  See more...
The millis setup code in wiring.c and wiring_lite.c has the almost by product of starting the timers for the analogWrite functions, without the millis setup, the timers never got started.

The setup, startup and shutdown of timers for analogWrite in particular is woefully complicated because wiring.c tries to work out how to do it for every chip.

To fix this, I have starting with tiny13 variant, setup a system of macros (or optional functions)

    turnOnPWM(t,v)
    turnOffPWM(t)

where t is the timer (eg TIMER0A) and v is the value 0-255

For tiny13, these are setup as simple macros because there is only 1 timer to worry about, it's easy and saves code space.

For other cores, they should make those defines call an actual function which they can create in their own pins_arduino.c, for example something like...

  #define turnOnPWM(t,v) ( _turnOnPWM(t,v) )
  void _turnOnPWM(uint8_t t, uint8_t v);

and create _turnOnPWM in your pins_arduino.c file to do whatever you need it to do to turn on a pwm pin for that timer with that value (which pin exactly is up to you if your hardware can put the timer on different ones).

That way ony the variant has to know how to turn it's pwm pins on/off, much much much easier to maintain!