Make optimisations to Print.h/cpp - tiny13 can now Serial.print with base 2/8/10/16 !
The original print class has a private method printNumber() which can print an `unsigned long` in an arbitrary base.

That's great and all, but it uses a TON of ram to do so for a few reasons.  Even in the best case it will use at least 37 bytes of stack, or more, probably more.  On a machine with 64 bytes of ram total like the t13 that's obviously unworkable.

In a previous commit I made an initial attempt to help, which redued flash usage but not SRAM, which is where the real problem was.

This rather large commit does it properly, in these ways.

  1. printNumber() has been replaced (optionally, but by default) with a ram (and flash too) efficient version which produces the same results
     (except in this commit it always returns 1... will fix that later) for **almost everybody**
  2. the maximum size of integer, that can be handled can be set by a define, by default it's `long` just like the original, but for small machines setting it to `int` or even `byte` is useful
  3. if the maximum integer is less than `long` then we get rid of some "casting" methods (for example print([int]) which just called print((long)[int]) )

by "almost everybody" I mean that everybody who wants to print a number in one of Base 2, Base 8, Base 10 or Base 16, will not find any difference.  Only if you want to print in some other peculiar base will you find it doesn't give you the result you wanted - in those cases, it will "promote" (or demote) your base to the higest supported one, base 16 by default, and it will kindly indicate this by printing "x" before the number (or b, o, or d depending on what the promotion/demotion went to).

For those who really, truely, desperately, need to print in an arbitrary base, pins_arduino.h can simply `#define PRINT_USE_BASE_ARBITRARY`, but come on, this is an ATtiny core, even big boys seldom need to use anything other than 2/8/10/16

Anyway, enough babble, see the comments in Print.h and tiny13/pins_arduino.h on how to use this all, remember by default no change is necessary and it should work just like it always did, as long as you only need base 2/8/10/16.

PS: ASCIITable works with Tiny13 now, including decimal output, provided you use NO_MILLIS, wrap the strings in F() and change `int thisByte` to `unsigned int thisByte` - frankly on a machine with 64bytes of RAM - (one-way) serial comms, string output, number conversion... that's pretty good going I think.