Add a "Serial Support" and "Print Support" to the boards menus for the Tiny13 which allows you to select different levels of support for each of those subsystems.
Notice how this has been done in platform.txt, it's not a **terrible** workaround for there being no ability to set globally applied `#define` from within a sketch.

Just means that if your sketch necessitates selecting some specific level of support (that is, to fit the sketch you have to sacrifice some print ability usually), that you'd better note it in a comment at the top of your sketch so you won't forget.