Change delayMicroseconds() to always be as "accurate" as we can, if millis() is enabled, then an adjustment is added for the interrupts that are likely to fire during this delay.  If millis() is not enabled then no adjustment is added.
Note that this is different to the Arduino mainline core, in that there they do not adjust for the interrupt handler at all, intending that the user will disable interrupts before calling delayMicroseconds(), but in practice this is probably not done and if you don't then the timing of a delay is **horribly** inaccurate.

delayMicroseconds is now a macro (alias) of either delayMicrosecondsWithoutMillisInterruptAdjustment, or delayMicrosecondsAdjustedForMillisInterrupt depending on if millis() is enabled or not.

Note that delayMicroseconds* is not affected by the millis accuracy, you can have 50% millis accuracy selected, and you will still get (reasonably) accurate delays out of delay() and delayMicroseconds().