Improve optimisation for pinMode,digitalRead,digitalWrite,analogRead,analogWrite
This commit makes these functions able to be much better (size) optimised for the most common use case where you are providing a constant pin number.  They also allow a pins_arduino.h to override the "default" implementation of these functions simply by defining their own as  a macro

    #define pinMode(p,m) myPinMode(p,m)

and implementing said myPinMode(p,m) in their pins_arduino.c (or directly in the macro, whatever).

These changes might look scary, but it's basically the same way that wiring does it these days I think, and I'm pretty confident that in the vast majority of cases we will find that the code space required with these optimisatons will be significantly lower than without.

There are a couple of "edge cases" where the optimisations could take more code space and there are optional defines to just make it fall-back to the normal method for those if you want - but I think it's unlikely to be a common thing to need to do:

  * #define OPTIMIZE_CONSTANT_PINMODE_INPUT 0
      Disables optimising pinMode(x,INPUT/INPUT_PULLUP)

  * #define OPTIMIZE_DIGITAL_WRITE_ON_TIMER 0
      Disables optimising digitalWrite(x,y) where x is on a timer pin

  * #define OPTIMIZE_DIGITAL_READ_ON_TIMER  0
      Disables optimising digitalRead(x) where x is on a timer pin

The optimisations in thise cases are not incorrect, they just might take a couple bytes more than the unoptimised version **if you do them more than a couple of times** (in which case the extra inline code taken would outweigh the overhead code of doing it in the default (non-inline) function.