Add support for flipped (inverted) analogWrite variant implementations.
The usual implementation of the analogWrite() timer setup is an 8 bit Fast PWM with the pin set at 0 and cleared at the desired value, that is, a desired value of "0" would in theory be fully off, and "255" be fully on.

However, this has a problem in that due to the way Fast PWM works, a desired value of 0 is not fully off, it's on for a very short time.  This is undesirable and the default implementation of analogWrite checks for 0 (and 255) and if those are the values uses digitalWrite() instead.

That means though that in order to use analogWrite(), digitalWrite() is also required to be linked by the compiler, which is really not nice if you don't need it.  It's almost always the case that the anlogWrite value is going to be a runtime known variable only, so we can't even do the static inlining trick, it won't help.

What we can do though is reverse the setup to be an 8 bit Fast PWM where the pin is cleared at 0 and set at 255.  This means that 0 is (almost) fully on, and 255 is fully off, something that is not quite fully-on (about 99.5% on) is typically indistinguisable from something that IS fully on, where as something that is not quite fully off is usually very apparant.  If we do this then we no longer need to check the values or use digitalWrite at all.

Naturally we don't want to confuse the user, so we take the usual value they give us as 0 (off) to 255 (on) and flip it before passing it into the AVR.