Add tone() support for Tiny13.
This is pretty laughable, but it works.  Since the t13 only has one Timer, it will mess with millis() while tones are playing, but due the lack of space you really can't have both millis() and tone().  Because division is so heavy, the calculations have been pulled out into an inlined function (in `tiny13/Tone.h`), so as long as you use a **constant** frequency and length in your call to tone() then no division is necessary and you can actually fit a few notes in (with millis disabled)....

    void loop()
    {
      tone(tonePin,660,100);
      delay(150);
      tone(tonePin,660,100);
      delay(300);
      tone(tonePin,660,100);
      delay(300);
      tone(tonePin,510,100);
      delay(100);
      tone(tonePin,660,100);
      delay(300);
      tone(tonePin,770,100);
      delay(550);
      tone(tonePin,380,100);
      delay(575);
      tone(tonePin,510,100);
      delay(450);
      tone(tonePin,380,100);
      delay(400);
      tone(tonePin,320,100);
      delay(500);
      tone(tonePin,440,100);
      delay(300);
      tone(tonePin,480,80);
      delay(330);
      tone(tonePin,450,100);
      delay(150);
      tone(tonePin,430,100);
      delay(300);
      tone(tonePin,380,100);
      delay(200);
      tone(tonePin,660,80);
      delay(200);
      tone(tonePin,760,50);
      delay(150);
      tone(tonePin,860,100);
      delay(300);
      tone(tonePin,700,80);
      delay(150);
      tone(tonePin,760,50);
      delay(350);
      tone(tonePin,660,80);
      delay(300);
      tone(tonePin,520,80);
      delay(150);
      tone(tonePin,580,80);
      delay(150);
    }