Remove pointless check.
There is only one pair of pin change interrupt registers, so there really is no need to be checking which one a pin is in.  If you give an invalid pin, that's your own fault, a Tiny processor doesn't have space to be sanity checking inputs (admittedly this is only an issue for non-constant pin numbers, but still).