First pass (untested, but compiles) of SPI for Tiny13
Taking the TinySoftwareSPI which was already existing in the core, stripping and optimising it down into a new library VeryTinySoftwareSPI which will be loaded by the (ATTinyCore) SPI library when that doesn't support the chip, namely in this case, for the ATTiny13 (and 4/5/9/10 but that's pretty unlikely to ever be a viable proposition).

Untested yet other than compilation of some simple SPI sketch (eg example for MCP41 digital pot) fits, fairly comfortably, in a Tiny13.

This probably isn't very useful in reality, but it's one of those because-I-can things.  Then again, I said the same about tone() on the Tiny13 and I use that quite often.